Funcionalidades de registro empleados

+++ Floristeria_SataUI/models/Employees.cs
@@ -8,11 +8,11 @@ namespace Floristeria_SataUI.models
 {
     public class Employees
     {
-        public int Documento { get; set; }
+        public long Documento { get; set; }
        public string Nombre { get; set; }
         public string Apellido { get; set; }
         public string Cargo { get; set; }
-        public int telefono { get; set; }
+        public long telefono { get; set; }
         public string imagen { get; set; }
